+++ cvat-canvas/src/canvasModel.ts
@@ -3,6 +3,9 @@
 * SPDX-License-Identifier: MIT
 */
 
+// Disable till full implementation
+/* eslint class-methods-use-this: "off" */
+
 import { MasterImpl } from './master';
 
 export interface Size {
@@ -202,7 +205,7 @@ export class CanvasModelImpl extends MasterImpl implements CanvasModel {
     }
 
     public draw(enabled: boolean, shapeType: string,
-                numberOfPoints: number, initialState: any): any {
+        numberOfPoints: number, initialState: any): any {
         return {
             enabled,
             initialState,