+++ cvat-core/jest.config.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018 Intel Corporation
+ * Copyright (C) 2019 Intel Corporation
  * SPDX-License-Identifier: MIT
  */